Updated examples and test script.

+++ src/shared/common.h
@@ -14,13 +14,13 @@
 
 using block_size_t = size_t;
 
-using class_t = uint64_t;
-using offset_t = uint64_t;
+using class_t = uint32_t;
+using offset_t = uint32_t;
 
-using block_bit_t = uint64_t;
+using block_bit_t = uint32_t;
 using rank_vector_per_bit_t = std::vector<class_t>;
 
-using bit_offset_t = uint64_t;
+using bit_offset_t = uint32_t;
 
 using block_vector_t = std::string;
